+++ crates/rendering/src/lib.rs
@@ -1,36 +1,38 @@
 use anyhow::Result;
 use bytemuck::{Pod, Zeroable};
-use cap_flags::FLAGS;
 use cap_project::{
     AspectRatio, BackgroundSource, CameraXPosition, CameraYPosition, Content, Crop,
-    CursorAnimationStyle, CursorClickEvent, CursorData, CursorEvents, CursorMoveEvent,
-    ProjectConfiguration, RecordingMeta, FAST_SMOOTHING_SAMPLES, FAST_VELOCITY_THRESHOLD,
-    REGULAR_SMOOTHING_SAMPLES, REGULAR_VELOCITY_THRESHOLD, SLOW_SMOOTHING_SAMPLES,
-    SLOW_VELOCITY_THRESHOLD, XY,
+    CursorAnimationStyle, CursorData, CursorEvents, CursorMoveEvent, ProjectConfiguration,
+    RecordingMeta, FAST_SMOOTHING_SAMPLES, FAST_VELOCITY_THRESHOLD, REGULAR_SMOOTHING_SAMPLES,
+    REGULAR_VELOCITY_THRESHOLD, SLOW_SMOOTHING_SAMPLES, SLOW_VELOCITY_THRESHOLD, XY,
 };
 use core::f64;
 use decoder::{spawn_decoder, AsyncVideoDecoderHandle};
+use frame_output::{FramePipelineEncoder, FramePipelineState};
 use futures::future::OptionFuture;
 use futures::FutureExt;
-use futures_intrusive::channel::shared::oneshot_channel;
-use reactive_graph::computed::ArcMemo;
-use reactive_graph::signal::{arc_signal, ArcReadSignal, ArcWriteSignal};
-use reactive_graph::traits::{Get, Write};
+use layers::{
+    BackgroundBlurPipeline, BackgroundLayer, CameraLayer, CursorLayer, CursorPipeline,
+    DisplayLayer, GradientOrColorPipeline, ImageBackgroundPipeline,
+};
 use serde::{Deserialize, Serialize};
 use specta::Type;
-use std::ops::{Add, Deref, Mul, Sub};
 use std::{collections::HashMap, sync::Arc};
 use tokio::sync::mpsc;
+use wgpu::include_wgsl;
 use wgpu::util::DeviceExt;
-use wgpu::{CommandEncoder, COPY_BYTES_PER_ROW_ALIGNMENT};
 
 use image::GenericImageView;
-use std::path::{Path, PathBuf};
-use std::time::{Duration, Instant};
+use std::{path::PathBuf, time::Instant};
 
+mod coord;
 pub mod decoder;
+mod frame_output;
+mod layers;
 mod project_recordings;
 mod zoom;
+
+pub use coord::*;
 pub use decoder::DecodedFrame;
 pub use project_recordings::{ProjectRecordings, SegmentRecordings, Video};
 
@@ -336,10 +338,6 @@ pub struct RenderVideoConstants {
     gradient_or_color_pipeline: GradientOrColorPipeline,
     pub cursor_textures: HashMap<String, wgpu::Texture>,
     cursor_pipeline: CursorPipeline,
-    watermark_pipeline: WatermarkPipeline,
-    watermark_texture: wgpu::Texture,
-    watermark_bind_group: wgpu::BindGroup,
-    watermark_dimensions: (u32, u32),
     image_background_pipeline: ImageBackgroundPipeline,
     background_textures: std::sync::Arc<tokio::sync::RwLock<HashMap<String, wgpu::Texture>>>,
     screen_frame: (wgpu::Texture, wgpu::TextureView),
@@ -367,95 +365,8 @@ impl RenderVideoConstants {
 
         let cursor_textures = Self::load_cursor_textures(&device, &queue, meta);
         let cursor_pipeline = CursorPipeline::new(&device);
-        let watermark_pipeline = WatermarkPipeline::new(&device);
         let composite_video_frame_pipeline = CompositeVideoFramePipeline::new(&device);
         let gradient_or_color_pipeline = GradientOrColorPipeline::new(&device);
-
-        let watermark_bytes = include_bytes!("../assets/watermark.png");
-        println!(
-            "Loading watermark texture from embedded bytes: {} bytes",
-            watermark_bytes.len()
-        );
-
-        let watermark_image = image::load_from_memory(watermark_bytes).unwrap();
-        let watermark_rgba = watermark_image.to_rgba8();
-        let dimensions = watermark_image.dimensions();
-        println!(
-            "Loaded watermark image dimensions: {}x{}",
-            dimensions.0, dimensions.1
-        );
-
-        let watermark_texture = device.create_texture(&wgpu::TextureDescriptor {
-            label: Some("WatermarkTexture"),
-            size: wgpu::Extent3d {
-                width: dimensions.0,
-                height: dimensions.1,
-                depth_or_array_layers: 1,
-            },
-            mip_level_count: 1,
-            sample_count: 1,
-            dimension: wgpu::TextureDimension::D2,
-            format: wgpu::TextureFormat::Rgba8UnormSrgb,
-            usage: wgpu::TextureUsages::TEXTURE_BINDING | wgpu::TextureUsages::COPY_DST,
-            view_formats: &[],
-        });
-
-        queue.write_texture(
-            wgpu::ImageCopyTexture {
-                texture: &watermark_texture,
-                mip_level: 0,
-                origin: wgpu::Origin3d::ZERO,
-                aspect: wgpu::TextureAspect::All,
-            },
-            &watermark_rgba,
-            wgpu::ImageDataLayout {
-                offset: 0,
-                bytes_per_row: Some(4 * dimensions.0),
-                rows_per_image: Some(dimensions.1),
-            },
-            wgpu::Extent3d {
-                width: dimensions.0,
-                height: dimensions.1,
-                depth_or_array_layers: 1,
-            },
-        );
-
-        let watermark_view = watermark_texture.create_view(&wgpu::TextureViewDescriptor::default());
-        let watermark_sampler = device.create_sampler(&wgpu::SamplerDescriptor {
-            address_mode_u: wgpu::AddressMode::ClampToEdge,
-            address_mode_v: wgpu::AddressMode::ClampToEdge,
-            address_mode_w: wgpu::AddressMode::ClampToEdge,
-            mag_filter: wgpu::FilterMode::Linear,
-            min_filter: wgpu::FilterMode::Linear,
-            mipmap_filter: wgpu::FilterMode::Linear,
-            ..Default::default()
-        });
-
-        let watermark_uniform_buffer = device.create_buffer(&wgpu::BufferDescriptor {
-            label: Some("WatermarkUniformBuffer"),
-            size: std::mem::size_of::<WatermarkUniforms>() as u64,
-            usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-            mapped_at_creation: false,
-        });
-
-        let watermark_bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            label: Some("WatermarkBindGroup"),
-            layout: &watermark_pipeline.bind_group_layout,
-            entries: &[
-                wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: wgpu::BindingResource::TextureView(&watermark_view),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 1,
-                    resource: wgpu::BindingResource::Sampler(&watermark_sampler),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 2,
-                    resource: watermark_uniform_buffer.as_entire_binding(),
-                },
-            ],
-        });
 
         let image_background_pipeline = ImageBackgroundPipeline::new(&device);
         let background_textures = Arc::new(tokio::sync::RwLock::new(HashMap::new()));
@@ -522,10 +433,6 @@ impl RenderVideoConstants {
             gradient_or_color_pipeline,
             cursor_textures,
             cursor_pipeline,
-            watermark_pipeline,
-            watermark_texture,
-            watermark_bind_group,
-            watermark_dimensions: dimensions,
             image_background_pipeline,
             background_textures,
             screen_frame,
@@ -1085,245 +992,16 @@ async fn produce_frame(
     resolution_base: XY<u32>,
     textures: &(wgpu::Texture, wgpu::Texture),
 ) -> Result<RenderedFrame, RenderingError> {
-    let mut encoder = constants.device.create_command_encoder(
-        &(wgpu::CommandEncoderDescriptor {
-            label: Some("Render Encoder"),
-        }),
-    );
+    let mut pipeline = FramePipelineState::new(constants, uniforms, textures);
+    let mut encoder = FramePipelineEncoder::new(&pipeline);
 
-    let textures = (&textures.0, &textures.1);
+    BackgroundLayer::new(&mut pipeline, &mut encoder)
+        .render(background)
+        .await?;
 
-    let texture_views = (
-        textures
-            .0
-            .create_view(&wgpu::TextureViewDescriptor::default()),
-        textures
-            .1
-            .create_view(&wgpu::TextureViewDescriptor::default()),
-    );
-    let texture_views = (&texture_views.0, &texture_views.1);
+    DisplayLayer::new(&mut pipeline, &mut encoder).render(&segment_frames);
 
-    let mut output_is_left = true;
-
-    // First, handle the background
-    match background {
-        Background::Image { path } => {
-            let mut textures = constants.background_textures.write().await;
-            let texture = match textures.entry(path.clone()) {
-                std::collections::hash_map::Entry::Occupied(e) => e.into_mut(),
-                std::collections::hash_map::Entry::Vacant(e) => {
-                    let img = image::open(&path)
-                        .map_err(|e| RenderingError::ImageLoadError(e.to_string()))?;
-                    let rgba = img.to_rgba8();
-                    let dimensions = img.dimensions();
-
-                    let texture = constants.device.create_texture(&wgpu::TextureDescriptor {
-                        label: Some("Background Image Texture"),
-                        size: wgpu::Extent3d {
-                            width: dimensions.0,
-                            height: dimensions.1,
-                            depth_or_array_layers: 1,
-                        },
-                        mip_level_count: 1,
-                        sample_count: 1,
-                        dimension: wgpu::TextureDimension::D2,
-                        format: wgpu::TextureFormat::Rgba8UnormSrgb,
-                        usage: wgpu::TextureUsages::TEXTURE_BINDING | wgpu::TextureUsages::COPY_DST,
-                        view_formats: &[],
-                    });
-
-                    constants.queue.write_texture(
-                        wgpu::ImageCopyTexture {
-                            texture: &texture,
-                            mip_level: 0,
-                            origin: wgpu::Origin3d::ZERO,
-                            aspect: wgpu::TextureAspect::All,
-                        },
-                        &rgba,
-                        wgpu::ImageDataLayout {
-                            offset: 0,
-                            bytes_per_row: Some(4 * dimensions.0),
-                            rows_per_image: Some(dimensions.1),
-                        },
-                        wgpu::Extent3d {
-                            width: dimensions.0,
-                            height: dimensions.1,
-                            depth_or_array_layers: 1,
-                        },
-                    );
-
-                    e.insert(texture)
-                }
-            };
-
-            let output_ar = uniforms.output_size.1 as f32 / uniforms.output_size.0 as f32;
-            let image_ar = texture.height() as f32 / texture.width() as f32;
-
-            let y_height = if output_ar < image_ar {
-                ((image_ar - output_ar) / 2.0) / image_ar
-            } else {
-                0.0
-            };
-
-            let x_width = if output_ar > image_ar {
-                let output_ar = 1.0 / output_ar;
-                let image_ar = 1.0 / image_ar;
-
-                ((image_ar - output_ar) / 2.0) / image_ar
-            } else {
-                0.0
-            };
-
-            let image_uniforms = ImageBackgroundUniforms {
-                output_size: [uniforms.output_size.0 as f32, uniforms.output_size.1 as f32],
-                padding: 0.0,
-                x_width,
-                y_height,
-                _padding: 0.0,
-            };
-
-            let uniform_buffer =
-                constants
-                    .device
-                    .create_buffer_init(&wgpu::util::BufferInitDescriptor {
-                        label: Some("Image Background Uniforms"),
-                        contents: bytemuck::cast_slice(&[image_uniforms]),
-                        usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-                    });
-
-            let texture_view = texture.create_view(&wgpu::TextureViewDescriptor::default());
-            let bind_group = constants.image_background_pipeline.bind_group(
-                &constants.device,
-                &uniform_buffer,
-                &texture_view,
-            );
-
-            do_render_pass(
-                &mut encoder,
-                get_either(texture_views, output_is_left),
-                &constants.image_background_pipeline.render_pipeline,
-                bind_group,
-                wgpu::LoadOp::Clear(wgpu::Color::BLACK),
-            );
-
-            output_is_left = !output_is_left;
-        }
-        _ => {
-            let bind_group = constants.gradient_or_color_pipeline.bind_group(
-                &constants.device,
-                &GradientOrColorUniforms::from(background).to_buffer(&constants.device),
-            );
-
-            do_render_pass(
-                &mut encoder,
-                get_either(texture_views, output_is_left),
-                &constants.gradient_or_color_pipeline.render_pipeline,
-                bind_group,
-                wgpu::LoadOp::Clear(wgpu::Color::BLACK),
-            );
-
-            output_is_left = !output_is_left;
-        }
-    }
-
-    {
-        let background_view = get_either(texture_views, !output_is_left);
-        if uniforms.project.background.blur > 0 {
-            let blur_strength = uniforms.project.background.blur as f32 / 100.0;
-            let blur_uniform = BackgroundBlurUniforms {
-                output_size: [uniforms.output_size.0 as f32, uniforms.output_size.1 as f32],
-                blur_strength,
-                _padding: 0.0,
-            };
-            let blur_buffer =
-                constants
-                    .device
-                    .create_buffer_init(&wgpu::util::BufferInitDescriptor {
-                        label: Some("BackgroundBlur Uniform Buffer"),
-                        contents: bytemuck::cast_slice(&[blur_uniform]),
-                        usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-                    });
-            let sampler = constants.device.create_sampler(&wgpu::SamplerDescriptor {
-                address_mode_u: wgpu::AddressMode::ClampToEdge,
-                address_mode_v: wgpu::AddressMode::ClampToEdge,
-                address_mode_w: wgpu::AddressMode::ClampToEdge,
-                mag_filter: wgpu::FilterMode::Linear,
-                min_filter: wgpu::FilterMode::Linear,
-                mipmap_filter: wgpu::FilterMode::Nearest,
-                ..Default::default()
-            });
-
-            // Use the other texture view as the output target
-            let output_view = get_either(texture_views, output_is_left);
-
-            let blur_bind_group = constants.background_blur_pipeline.bind_group(
-                &constants.device,
-                &blur_buffer,
-                background_view,
-                &sampler,
-            );
-
-            do_render_pass(
-                &mut encoder,
-                output_view,
-                &constants.background_blur_pipeline.render_pipeline,
-                blur_bind_group,
-                wgpu::LoadOp::Clear(wgpu::Color::BLACK),
-            );
-
-            output_is_left = !output_is_left;
-        }
-    }
-
-    // Then render the screen frame
-    {
-        let frame_size = constants.options.screen_size;
-
-        constants.queue.write_texture(
-            wgpu::ImageCopyTexture {
-                texture: &constants.screen_frame.0,
-                mip_level: 0,
-                origin: wgpu::Origin3d::ZERO,
-                aspect: wgpu::TextureAspect::All,
-            },
-            &segment_frames.screen_frame,
-            wgpu::ImageDataLayout {
-                offset: 0,
-                bytes_per_row: Some(constants.options.screen_size.x * 4),
-                rows_per_image: None,
-            },
-            wgpu::Extent3d {
-                width: frame_size.x,
-                height: frame_size.y,
-                depth_or_array_layers: 1,
-            },
-        );
-
-        do_render_pass(
-            &mut encoder,
-            get_either(texture_views, output_is_left),
-            &constants.composite_video_frame_pipeline.render_pipeline,
-            constants.composite_video_frame_pipeline.bind_group(
-                &constants.device,
-                &uniforms.display.to_buffer(&constants.device),
-                &constants.screen_frame.1,
-                get_either(texture_views, !output_is_left),
-            ),
-            wgpu::LoadOp::Load,
-        );
-
-        output_is_left = !output_is_left;
-    }
-
-    // Then render the cursor
-    draw_cursor(
-        constants,
-        uniforms,
-        segment_frames.segment_time,
-        &mut encoder,
-        get_either(texture_views, !output_is_left),
-        resolution_base,
-    );
+    CursorLayer::new(&mut pipeline, &mut encoder).render(&segment_frames, resolution_base);
 
     // camera
     if let (Some(camera_size), Some(camera_frame), Some(uniforms), Some((texture, texture_view))) = (
@@ -1332,222 +1010,16 @@ async fn produce_frame(
         &uniforms.camera,
         &constants.camera_frame,
     ) {
-        constants.queue.write_texture(
-            wgpu::ImageCopyTexture {
-                texture: &texture,
-                mip_level: 0,
-                origin: wgpu::Origin3d::ZERO,
-                aspect: wgpu::TextureAspect::All,
-            },
+        CameraLayer::new(&mut pipeline, &mut encoder).render(
+            camera_size,
             camera_frame,
-            wgpu::ImageDataLayout {
-                offset: 0,
-                bytes_per_row: Some(camera_size.x * 4),
-                rows_per_image: None,
-            },
-            wgpu::Extent3d {
-                width: camera_size.x,
-                height: camera_size.y,
-                depth_or_array_layers: 1,
-            },
+            uniforms,
+            (texture, texture_view),
         );
-
-        do_render_pass(
-            &mut encoder,
-            get_either(texture_views, output_is_left),
-            &constants.composite_video_frame_pipeline.render_pipeline,
-            constants.composite_video_frame_pipeline.bind_group(
-                &constants.device,
-                &uniforms.to_buffer(&constants.device),
-                &texture_view,
-                get_either(texture_views, !output_is_left),
-            ),
-            wgpu::LoadOp::Load, // Load existing content
-        );
-
-        output_is_left = !output_is_left;
     }
 
-    // Now submit the encoder
-    constants.queue.submit(std::iter::once(encoder.finish()));
-
-    // Add watermark if not upgraded - moved to after all other rendering
-    if uniforms.project.background.inset <= 0 {
-        // Calculate watermark width as 10% of frame width, but no larger than 200px
-        let frame_width = uniforms.output_size.0 as f32;
-        let target_width = (frame_width * 0.15).min(280.0);
-        let aspect_ratio =
-            constants.watermark_dimensions.1 as f32 / constants.watermark_dimensions.0 as f32;
-
-        // Increase padding from edges
-        let padding = frame_width * 0.02; // 2% of frame width for padding
-
-        let watermark_uniforms = WatermarkUniforms {
-            output_size: [uniforms.output_size.0 as f32, uniforms.output_size.1 as f32],
-            watermark_size: [target_width, target_width * aspect_ratio],
-            position: [
-                padding,
-                uniforms.output_size.1 as f32 - (target_width * aspect_ratio) - padding,
-            ],
-            opacity: if uniforms.is_upgraded { 0.0 } else { 0.5 },
-            is_upgraded: if uniforms.is_upgraded { 1.0 } else { 0.0 },
-        };
-
-        // println!(
-        //     "Rendering watermark at position: {:?}, size: {:?}, output_size: {:?}",
-        //     watermark_uniforms.position,
-        //     watermark_uniforms.watermark_size,
-        //     watermark_uniforms.output_size
-        // );
-
-        let watermark_buffer =
-            constants
-                .device
-                .create_buffer_init(&wgpu::util::BufferInitDescriptor {
-                    label: Some("Watermark Uniforms Buffer"),
-                    contents: bytemuck::cast_slice(&[watermark_uniforms]),
-                    usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-                });
-
-        let watermark_bind_group = constants
-            .device
-            .create_bind_group(&wgpu::BindGroupDescriptor {
-                label: Some("WatermarkBindGroup"),
-                layout: &constants.watermark_pipeline.bind_group_layout,
-                entries: &[
-                    wgpu::BindGroupEntry {
-                        binding: 0,
-                        resource: wgpu::BindingResource::TextureView(
-                            &constants
-                                .watermark_texture
-                                .create_view(&wgpu::TextureViewDescriptor::default()),
-                        ),
-                    },
-                    wgpu::BindGroupEntry {
-                        binding: 1,
-                        resource: wgpu::BindingResource::Sampler(&constants.device.create_sampler(
-                            &wgpu::SamplerDescriptor {
-                                address_mode_u: wgpu::AddressMode::ClampToEdge,
-                                address_mode_v: wgpu::AddressMode::ClampToEdge,
-                                address_mode_w: wgpu::AddressMode::ClampToEdge,
-                                mag_filter: wgpu::FilterMode::Linear,
-                                min_filter: wgpu::FilterMode::Linear,
-                                mipmap_filter: wgpu::FilterMode::Linear,
-                                anisotropy_clamp: 16, // Add anisotropic filtering
-                                ..Default::default()
-                            },
-                        )),
-                    },
-                    wgpu::BindGroupEntry {
-                        binding: 2,
-                        resource: watermark_buffer.as_entire_binding(),
-                    },
-                ],
-            });
-
-        let mut watermark_encoder =
-            constants
-                .device
-                .create_command_encoder(&wgpu::CommandEncoderDescriptor {
-                    label: Some("WatermarkEncoder"),
-                });
-
-        {
-            let mut render_pass =
-                watermark_encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-                    label: Some("WatermarkRenderPass"),
-                    color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                        view: get_either(texture_views, !output_is_left),
-                        resolve_target: None,
-                        ops: wgpu::Operations {
-                            load: wgpu::LoadOp::Load,
-                            store: wgpu::StoreOp::Store,
-                        },
-                    })],
-                    depth_stencil_attachment: None,
-                    timestamp_writes: None,
-                    occlusion_query_set: None,
-                });
-
-            render_pass.set_pipeline(&constants.watermark_pipeline.render_pipeline);
-            render_pass.set_bind_group(0, &watermark_bind_group, &[]);
-            render_pass.draw(0..6, 0..1); // Using 6 vertices for two triangles
-                                          // println!("Drew watermark with 6 vertices");
-        }
-
-        constants.queue.submit(Some(watermark_encoder.finish()));
-    }
-
-    let output_texture_size = wgpu::Extent3d {
-        width: uniforms.output_size.0,
-        height: uniforms.output_size.1,
-        depth_or_array_layers: 1,
-    };
-
-    // Calculate the aligned bytes per row
-    let align = COPY_BYTES_PER_ROW_ALIGNMENT;
-    let unpadded_bytes_per_row = uniforms.output_size.0 * 4;
-    let padding = (align - (unpadded_bytes_per_row % align)) % align;
-    let padded_bytes_per_row = unpadded_bytes_per_row + padding;
-
-    // Ensure the padded_bytes_per_row is a multiple of 4 (32 bits)
-    let padded_bytes_per_row = (padded_bytes_per_row + 3) & !3;
-
-    let output_buffer_size = (padded_bytes_per_row * uniforms.output_size.1) as u64;
-
-    let output_buffer = constants.device.create_buffer(&wgpu::BufferDescriptor {
-        size: output_buffer_size,
-        usage: wgpu::BufferUsages::COPY_DST | wgpu::BufferUsages::MAP_READ,
-        label: Some("Output Buffer"),
-        mapped_at_creation: false,
-    });
-
-    {
-        let mut encoder = constants.device.create_command_encoder(
-            &(wgpu::CommandEncoderDescriptor {
-                label: Some("Copy Encoder"),
-            }),
-        );
-
-        encoder.copy_texture_to_buffer(
-            wgpu::ImageCopyTexture {
-                texture: get_either(textures, !output_is_left),
-                mip_level: 0,
-                origin: wgpu::Origin3d::ZERO,
-                aspect: wgpu::TextureAspect::All,
-            },
-            wgpu::ImageCopyBuffer {
-                buffer: &output_buffer,
-                layout: wgpu::ImageDataLayout {
-                    offset: 0,
-                    bytes_per_row: Some(padded_bytes_per_row),
-                    rows_per_image: Some(uniforms.output_size.1),
-                },
-            },
-            output_texture_size,
-        );
-
-        constants.queue.submit(std::iter::once(encoder.finish()));
-    }
-
-    let buffer_slice = output_buffer.slice(..);
-    let (tx, rx) = oneshot_channel();
-    buffer_slice.map_async(wgpu::MapMode::Read, move |result| {
-        tx.send(result).ok();
-    });
-    constants.device.poll(wgpu::Maintain::Wait);
-
-    rx.receive()
-        .await
-        .ok_or(RenderingError::BufferMapWaitingFailed)??;
-
-    let data = buffer_slice.get_mapped_range();
-
-    let image_data = data.to_vec();
-
-    // Unmap the buffer
-    drop(data);
-    output_buffer.unmap();
+    let padded_bytes_per_row = encoder.padded_bytes_per_row(&pipeline);
+    let image_data = encoder.copy_output(pipeline).await?;
 
     Ok(RenderedFrame {
         data: image_data,
@@ -1555,156 +1027,6 @@ async fn produce_frame(
         width: uniforms.output_size.0,
         height: uniforms.output_size.1,
     })
-}
-
-fn draw_cursor(
-    constants: &RenderVideoConstants,
-    uniforms: &ProjectUniforms,
-    segment_time: f32,
-    encoder: &mut CommandEncoder,
-    view: &wgpu::TextureView,
-    resolution_base: XY<u32>,
-) {
-    let Some(cursor_position) = interpolate_cursor_position(
-        &Default::default(), // constants.cursor,
-        segment_time,
-        &uniforms.project.cursor.animation_style,
-    ) else {
-        return;
-    };
-
-    // Calculate previous position for velocity
-    let prev_position = interpolate_cursor_position(
-        &Default::default(), // constants.cursor,
-        segment_time - 1.0 / 30.0,
-        &uniforms.project.cursor.animation_style,
-    );
-
-    // Calculate velocity in screen space
-    let velocity = if let Some(prev_pos) = prev_position {
-        let curr_frame_pos =
-            cursor_position.to_frame_space(&constants.options, &uniforms.project, resolution_base);
-        let prev_frame_pos =
-            prev_pos.to_frame_space(&constants.options, &uniforms.project, resolution_base);
-        let frame_velocity = curr_frame_pos.coord - prev_frame_pos.coord;
-
-        // Convert to pixels per frame
-        [frame_velocity.x as f32, frame_velocity.y as f32]
-    } else {
-        [0.0, 0.0]
-    };
-
-    // Calculate motion blur amount based on velocity magnitude
-    let speed = (velocity[0] * velocity[0] + velocity[1] * velocity[1]).sqrt();
-    let motion_blur_amount = (speed * 0.3).min(1.0) * uniforms.project.motion_blur.unwrap_or(0.8);
-
-    let cursor = Default::default();
-    let cursor_event = find_cursor_event(&cursor /* constants.cursor */, segment_time);
-
-    let last_click_time =  /* constants
-        .cursor
-        .clicks */ Vec::<CursorClickEvent>::new()
-        .iter()
-        .filter(|click| click.down && click.process_time_ms <= (segment_time as f64) * 1000.0)
-        .max_by_key(|click| click.process_time_ms as i64)
-        .map(|click| ((segment_time as f64) * 1000.0 - click.process_time_ms) as f32 / 1000.0)
-        .unwrap_or(1.0);
-
-    let Some(cursor_texture) = constants.cursor_textures.get(&cursor_event.cursor_id) else {
-        return;
-    };
-
-    let cursor_size = cursor_texture.size();
-    let aspect_ratio = cursor_size.width as f32 / cursor_size.height as f32;
-
-    let cursor_size_percentage = if uniforms.cursor_size <= 0.0 {
-        100.0
-    } else {
-        uniforms.cursor_size / 100.0
-    };
-
-    let normalized_size = [
-        STANDARD_CURSOR_HEIGHT * aspect_ratio * cursor_size_percentage,
-        STANDARD_CURSOR_HEIGHT * cursor_size_percentage,
-    ];
-
-    let frame_position =
-        cursor_position.to_frame_space(&constants.options, &uniforms.project, resolution_base);
-    // let position = uniforms.zoom.apply_scale(frame_position);
-    let position = frame_position;
-    let relative_position = [position.x as f32, position.y as f32];
-
-    let cursor_uniforms = CursorUniforms {
-        position: [relative_position[0], relative_position[1], 0.0, 0.0],
-        size: [normalized_size[0], normalized_size[1], 0.0, 0.0],
-        output_size: [
-            uniforms.output_size.0 as f32,
-            uniforms.output_size.1 as f32,
-            0.0,
-            0.0,
-        ],
-        screen_bounds: uniforms.display.target_bounds,
-        cursor_size: cursor_size_percentage,
-        last_click_time,
-        velocity,
-        motion_blur_amount,
-        _alignment: [0.0; 7],
-    };
-
-    let cursor_uniform_buffer =
-        constants
-            .device
-            .create_buffer_init(&wgpu::util::BufferInitDescriptor {
-                label: Some("Cursor Uniform Buffer"),
-                contents: bytemuck::cast_slice(&[cursor_uniforms]),
-                usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-            });
-
-    let cursor_bind_group = constants
-        .device
-        .create_bind_group(&wgpu::BindGroupDescriptor {
-            layout: &constants.cursor_pipeline.bind_group_layout,
-            entries: &[
-                wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: cursor_uniform_buffer.as_entire_binding(),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 1,
-                    resource: wgpu::BindingResource::TextureView(
-                        &cursor_texture.create_view(&wgpu::TextureViewDescriptor::default()),
-                    ),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 2,
-                    resource: wgpu::BindingResource::Sampler(
-                        &constants
-                            .device
-                            .create_sampler(&wgpu::SamplerDescriptor::default()),
-                    ),
-                },
-            ],
-            label: Some("Cursor Bind Group"),
-        });
-
-    let mut render_pass = encoder.begin_render_pass(&wgpu::RenderPassDescriptor {
-        label: Some("Cursor Render Pass"),
-        color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-            view,
-            resolve_target: None,
-            ops: wgpu::Operations {
-                load: wgpu::LoadOp::Load,
-                store: wgpu::StoreOp::Store,
-            },
-        })],
-        depth_stencil_attachment: None,
-        timestamp_writes: None,
-        occlusion_query_set: None,
-    });
-
-    render_pass.set_pipeline(&constants.cursor_pipeline.render_pipeline);
-    render_pass.set_bind_group(0, &cursor_bind_group, &[]);
-    render_pass.draw(0..4, 0..1);
 }
 
 struct CompositeVideoFramePipeline {
@@ -1747,17 +1069,16 @@ impl CompositeVideoFrameUniforms {
 impl CompositeVideoFramePipeline {
     pub fn new(device: &wgpu::Device) -> Self {
         let bind_group_layout = Self::bind_group_layout(device);
-        let render_pipeline =
-            create_shader_render_pipeline(device, &bind_group_layout, Self::shader());
+        let render_pipeline = create_shader_render_pipeline(
+            device,
+            &bind_group_layout,
+            include_wgsl!("shaders/composite-video-frame.wgsl"),
+        );
 
         Self {
             bind_group_layout,
             render_pipeline,
         }
-    }
-
-    fn shader() -> &'static str {
-        include_str!("shaders/composite-video-frame.wgsl")
     }
 
     fn bind_group_layout(device: &wgpu::Device) -> wgpu::BindGroupLayout {
@@ -1852,269 +1173,12 @@ impl CompositeVideoFramePipeline {
     }
 }
 
-struct GradientOrColorPipeline {
-    pub bind_group_layout: wgpu::BindGroupLayout,
-    pub render_pipeline: wgpu::RenderPipeline,
-}
-
-#[derive(Debug, Clone, Copy, Pod, Zeroable, Default)]
-#[repr(C)]
-struct GradientOrColorUniforms {
-    pub start: [f32; 4],
-    pub end: [f32; 4],
-    pub angle: f32,
-    _padding: [f32; 3],
-}
-
-impl GradientOrColorUniforms {
-    fn to_buffer(self, device: &wgpu::Device) -> wgpu::Buffer {
-        device.create_buffer_init(
-            &(wgpu::util::BufferInitDescriptor {
-                label: Some("GradientOrColorUniforms Buffer"),
-                contents: bytemuck::cast_slice(&[self]),
-                usage: wgpu::BufferUsages::UNIFORM | wgpu::BufferUsages::COPY_DST,
-            }),
-        )
-    }
-}
-
-impl From<Background> for GradientOrColorUniforms {
-    fn from(value: Background) -> Self {
-        match value {
-            Background::Color(color) => Self {
-                start: color,
-                end: color,
-                angle: 0.0,
-                _padding: [0.0; 3],
-            },
-            Background::Gradient { start, end, angle } => Self {
-                start,
-                end,
-                angle,
-                _padding: [0.0; 3],
-            },
-            Background::Image { .. } => {
-                unreachable!("Image backgrounds should be handled separately")
-            }
-        }
-    }
-}
-
-impl GradientOrColorPipeline {
-    pub fn new(device: &wgpu::Device) -> Self {
-        let bind_group_layout = Self::bind_group_layout(device);
-        let render_pipeline =
-            create_shader_render_pipeline(device, &bind_group_layout, Self::shader());
-
-        Self {
-            bind_group_layout,
-            render_pipeline,
-        }
-    }
-
-    fn shader() -> &'static str {
-        include_str!("shaders/gradient-or-color.wgsl")
-    }
-
-    fn bind_group_layout(device: &wgpu::Device) -> wgpu::BindGroupLayout {
-        device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-            label: Some("composite-video-frame.wgsl Bind Group Layout"),
-            entries: &[wgpu::BindGroupLayoutEntry {
-                binding: 0,
-                visibility: wgpu::ShaderStages::VERTEX_FRAGMENT,
-                ty: wgpu::BindingType::Buffer {
-                    ty: wgpu::BufferBindingType::Uniform,
-                    has_dynamic_offset: false,
-                    min_binding_size: None,
-                },
-                count: None,
-            }],
-        })
-    }
-
-    pub fn bind_group(&self, device: &wgpu::Device, uniforms: &wgpu::Buffer) -> wgpu::BindGroup {
-        let bind_group = device.create_bind_group(&wgpu::BindGroupDescriptor {
-            layout: &self.bind_group_layout,
-            entries: &[wgpu::BindGroupEntry {
-                binding: 0,
-                resource: uniforms.as_entire_binding(),
-            }],
-            label: Some("bind_group"),
-        });
-
-        bind_group
-    }
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Pod, Zeroable)]
-struct BackgroundBlurUniforms {
-    output_size: [f32; 2],
-    blur_strength: f32,
-    _padding: f32,
-}
-
-struct BackgroundBlurPipeline {
-    pub bind_group_layout: wgpu::BindGroupLayout,
-    pub render_pipeline: wgpu::RenderPipeline,
-}
-
-impl BackgroundBlurPipeline {
-    pub fn new(device: &wgpu::Device) -> Self {
-        let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-            label: Some("background-blur Bind Group Layout"),
-            entries: &[
-                wgpu::BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: wgpu::ShaderStages::VERTEX_FRAGMENT,
-                    ty: wgpu::BindingType::Buffer {
-                        ty: wgpu::BufferBindingType::Uniform,
-                        has_dynamic_offset: false,
-                        min_binding_size: None,
-                    },
-                    count: None,
-                },
-                wgpu::BindGroupLayoutEntry {
-                    binding: 1,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Texture {
-                        sample_type: wgpu::TextureSampleType::Float { filterable: true },
-                        view_dimension: wgpu::TextureViewDimension::D2,
-                        multisampled: false,
-                    },
-                    count: None,
-                },
-                wgpu::BindGroupLayoutEntry {
-                    binding: 2,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::Filtering),
-                    count: None,
-                },
-            ],
-        });
-        let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
-            label: Some("Background Blur Shader"),
-            source: wgpu::ShaderSource::Wgsl(include_str!("shaders/background-blur.wgsl").into()),
-        });
-        let empty_constants: HashMap<String, f64> = HashMap::new();
-        let pipeline_layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-            label: Some("Background Blur Pipeline Layout"),
-            bind_group_layouts: &[&bind_group_layout],
-            push_constant_ranges: &[],
-        });
-        let render_pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-            label: Some("Background Blur Pipeline"),
-            layout: Some(&pipeline_layout),
-            vertex: wgpu::VertexState {
-                module: &shader,
-                entry_point: "vs_main",
-                buffers: &[],
-                compilation_options: wgpu::PipelineCompilationOptions {
-                    constants: &empty_constants,
-                    zero_initialize_workgroup_memory: false,
-                    vertex_pulling_transform: false,
-                },
-            },
-            fragment: Some(wgpu::FragmentState {
-                module: &shader,
-                entry_point: "fs_main",
-                targets: &[Some(wgpu::ColorTargetState {
-                    format: wgpu::TextureFormat::Rgba8UnormSrgb,
-                    blend: Some(wgpu::BlendState::REPLACE),
-                    write_mask: wgpu::ColorWrites::ALL,
-                })],
-                compilation_options: wgpu::PipelineCompilationOptions {
-                    constants: &empty_constants,
-                    zero_initialize_workgroup_memory: false,
-                    vertex_pulling_transform: false,
-                },
-            }),
-            primitive: wgpu::PrimitiveState {
-                topology: wgpu::PrimitiveTopology::TriangleList,
-                strip_index_format: None,
-                front_face: wgpu::FrontFace::Ccw,
-                cull_mode: Some(wgpu::Face::Back),
-                polygon_mode: wgpu::PolygonMode::Fill,
-                unclipped_depth: false,
-                conservative: false,
-            },
-            depth_stencil: None,
-            multisample: wgpu::MultisampleState::default(),
-            multiview: None,
-            cache: None,
-        });
-        Self {
-            bind_group_layout,
-            render_pipeline,
-        }
-    }
-
-    pub fn bind_group(
-        &self,
-        device: &wgpu::Device,
-        uniform_buffer: &wgpu::Buffer,
-        texture_view: &wgpu::TextureView,
-        sampler: &wgpu::Sampler,
-    ) -> wgpu::BindGroup {
-        device.create_bind_group(&wgpu::BindGroupDescriptor {
-            label: Some("BackgroundBlur Bind Group"),
-            layout: &self.bind_group_layout,
-            entries: &[
-                wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: uniform_buffer.as_entire_binding(),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 1,
-                    resource: wgpu::BindingResource::TextureView(texture_view),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 2,
-                    resource: wgpu::BindingResource::Sampler(sampler),
-                },
-            ],
-        })
-    }
-}
-
-fn do_render_pass(
-    encoder: &mut wgpu::CommandEncoder,
-    output_view: &wgpu::TextureView,
-    render_pipeline: &wgpu::RenderPipeline,
-    bind_group: wgpu::BindGroup,
-    load_op: wgpu::LoadOp<wgpu::Color>,
-) {
-    let mut render_pass = encoder.begin_render_pass(
-        &(wgpu::RenderPassDescriptor {
-            label: Some("Render Pass"),
-            color_attachments: &[Some(wgpu::RenderPassColorAttachment {
-                view: output_view,
-                resolve_target: None,
-                ops: wgpu::Operations {
-                    load: load_op,
-                    store: wgpu::StoreOp::Store,
-                },
-            })],
-            depth_stencil_attachment: None,
-            timestamp_writes: None,
-            occlusion_query_set: None,
-        }),
-    );
-
-    render_pass.set_pipeline(render_pipeline);
-    render_pass.set_bind_group(0, &bind_group, &[]);
-    render_pass.draw(0..3, 0..1);
-}
-
-fn create_shader_render_pipeline(
+pub fn create_shader_render_pipeline(
     device: &wgpu::Device,
     bind_group_layout: &wgpu::BindGroupLayout,
-    wgsl_shader: &str,
+    shader: wgpu::ShaderModuleDescriptor,
 ) -> wgpu::RenderPipeline {
-    let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
-        label: Some("Shader"),
-        source: wgpu::ShaderSource::Wgsl(wgsl_shader.into()),
-    });
+    let shader = device.create_shader_module(shader);
 
     let pipeline_layout = device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
         label: Some("Render Pipeline Layout"),
@@ -2303,537 +1367,4 @@ fn interpolate_cursor_position(
         x: x.clamp(0.0, 1.0),
         y: y.clamp(0.0, 1.0),
     }))
-}
-
-struct CursorPipeline {
-    bind_group_layout: wgpu::BindGroupLayout,
-    render_pipeline: wgpu::RenderPipeline,
-}
-
-#[repr(C, align(16))]
-#[derive(Debug, Clone, Copy, Pod, Zeroable)]
-struct CursorUniforms {
-    position: [f32; 4],
-    size: [f32; 4],
-    output_size: [f32; 4],
-    screen_bounds: [f32; 4],
-    cursor_size: f32,
-    last_click_time: f32,
-    velocity: [f32; 2],
-    motion_blur_amount: f32,
-    _alignment: [f32; 7],
-}
-
-fn find_cursor_event(cursor: &CursorData, time: f32) -> &CursorMoveEvent {
-    let time_ms = time * 1000.0;
-
-    let event = cursor
-        .moves
-        .iter()
-        .rev()
-        .find(|event| {
-            // println!("Checking event at time: {}ms", event.process_time_ms);
-            event.process_time_ms <= time_ms.into()
-        })
-        .unwrap_or(&cursor.moves[0]);
-
-    event
-}
-
-impl CursorPipeline {
-    pub fn new(device: &wgpu::Device) -> Self {
-        let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-            label: Some("Cursor Pipeline Layout"),
-            entries: &[
-                wgpu::BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: wgpu::ShaderStages::VERTEX_FRAGMENT,
-                    ty: wgpu::BindingType::Buffer {
-                        ty: wgpu::BufferBindingType::Uniform,
-                        has_dynamic_offset: false,
-                        min_binding_size: Some(std::num::NonZeroU64::new(112).unwrap()),
-                    },
-                    count: None,
-                },
-                wgpu::BindGroupLayoutEntry {
-                    binding: 1,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Texture {
-                        sample_type: wgpu::TextureSampleType::Float { filterable: true },
-                        view_dimension: wgpu::TextureViewDimension::D2,
-                        multisampled: false,
-                    },
-                    count: None,
-                },
-                wgpu::BindGroupLayoutEntry {
-                    binding: 2,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::Filtering),
-                    count: None,
-                },
-            ],
-        });
-
-        let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
-            label: Some("Cursor Shader"),
-            source: wgpu::ShaderSource::Wgsl(include_str!("shaders/cursor.wgsl").into()),
-        });
-
-        let empty_constants: HashMap<String, f64> = HashMap::new();
-
-        let render_pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-            label: Some("Cursor Pipeline"),
-            layout: Some(
-                &device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-                    label: Some("Cursor Pipeline Layout"),
-                    bind_group_layouts: &[&bind_group_layout],
-                    push_constant_ranges: &[],
-                }),
-            ),
-            vertex: wgpu::VertexState {
-                module: &shader,
-                entry_point: "vs_main",
-                buffers: &[],
-                compilation_options: wgpu::PipelineCompilationOptions {
-                    constants: &empty_constants,
-                    zero_initialize_workgroup_memory: false,
-                    vertex_pulling_transform: false,
-                },
-            },
-            fragment: Some(wgpu::FragmentState {
-                module: &shader,
-                entry_point: "fs_main",
-                targets: &[Some(wgpu::ColorTargetState {
-                    format: wgpu::TextureFormat::Rgba8UnormSrgb,
-                    blend: Some(wgpu::BlendState {
-                        color: wgpu::BlendComponent {
-                            src_factor: wgpu::BlendFactor::One,
-                            dst_factor: wgpu::BlendFactor::OneMinusSrcAlpha,
-                            operation: wgpu::BlendOperation::Add,
-                        },
-                        alpha: wgpu::BlendComponent {
-                            src_factor: wgpu::BlendFactor::One,
-                            dst_factor: wgpu::BlendFactor::OneMinusSrcAlpha,
-                            operation: wgpu::BlendOperation::Add,
-                        },
-                    }),
-                    write_mask: wgpu::ColorWrites::ALL,
-                })],
-                compilation_options: wgpu::PipelineCompilationOptions {
-                    constants: &empty_constants,
-                    zero_initialize_workgroup_memory: false,
-                    vertex_pulling_transform: false,
-                },
-            }),
-            primitive: wgpu::PrimitiveState {
-                topology: wgpu::PrimitiveTopology::TriangleStrip,
-                strip_index_format: None,
-                front_face: wgpu::FrontFace::Ccw,
-                cull_mode: None,
-                unclipped_depth: false,
-                polygon_mode: wgpu::PolygonMode::Fill,
-                conservative: false,
-            },
-            depth_stencil: None,
-            multisample: wgpu::MultisampleState::default(),
-            multiview: None,
-            cache: None,
-        });
-
-        Self {
-            bind_group_layout,
-            render_pipeline,
-        }
-    }
-}
-
-#[derive(Default, Clone, Copy, Debug)]
-struct RawDisplaySpace;
-
-// raw cursor data
-#[derive(Default, Clone, Copy, Debug)]
-struct RawDisplayUVSpace;
-
-#[derive(Default, Clone, Copy, Debug)]
-struct CroppedDisplaySpace;
-
-#[derive(Default, Clone, Copy, Debug)]
-pub struct FrameSpace;
-
-#[derive(Default, Clone, Copy, Debug)]
-struct TransformedDisplaySpace;
-
-#[derive(Clone, Copy, Debug)]
-pub struct Coord<TSpace> {
-    coord: XY<f64>,
-    space: TSpace,
-}
-
-impl<TSpace: Default> Coord<TSpace> {
-    pub fn new(coord: XY<f64>) -> Self {
-        Self {
-            coord,
-            space: TSpace::default(),
-        }
-    }
-
-    pub fn clamp(self, min: XY<f64>, max: XY<f64>) -> Self {
-        Self {
-            coord: XY {
-                x: self.coord.x.clamp(min.x, max.x),
-                y: self.coord.y.clamp(min.y, max.y),
-            },
-            space: self.space,
-        }
-    }
-}
-
-impl<T> Deref for Coord<T> {
-    type Target = XY<f64>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.coord
-    }
-}
-
-impl Coord<RawDisplayUVSpace> {
-    fn to_raw_display_space(&self, options: &RenderOptions) -> Coord<RawDisplaySpace> {
-        Coord::new(self.coord * options.screen_size.map(|v| v as f64))
-    }
-
-    fn to_frame_space(
-        &self,
-        options: &RenderOptions,
-        project: &ProjectConfiguration,
-        resolution_base: XY<u32>,
-    ) -> Coord<FrameSpace> {
-        self.to_raw_display_space(options)
-            .to_cropped_display_space(options, project)
-            .to_frame_space(options, project, resolution_base)
-    }
-}
-
-impl Coord<RawDisplaySpace> {
-    fn to_cropped_display_space(
-        &self,
-        options: &RenderOptions,
-        project: &ProjectConfiguration,
-    ) -> Coord<CroppedDisplaySpace> {
-        let crop = ProjectUniforms::get_crop(options, project);
-        Coord::new(self.coord - crop.position.map(|v| v as f64))
-    }
-}
-
-impl Coord<CroppedDisplaySpace> {
-    fn to_frame_space(
-        &self,
-        options: &RenderOptions,
-        project: &ProjectConfiguration,
-        resolution_base: XY<u32>,
-    ) -> Coord<FrameSpace> {
-        let padding = ProjectUniforms::get_display_offset(options, project, resolution_base);
-        Coord::new(self.coord + *padding)
-    }
-}
-
-impl<T> Add for Coord<T> {
-    type Output = Self;
-
-    fn add(self, rhs: Self) -> Self {
-        Coord {
-            coord: self.coord + rhs.coord,
-            space: self.space,
-        }
-    }
-}
-
-impl<T> Sub for Coord<T> {
-    type Output = Self;
-
-    fn sub(self, rhs: Self) -> Self {
-        Coord {
-            coord: self.coord - rhs.coord,
-            space: self.space,
-        }
-    }
-}
-
-impl<T> Mul<f64> for Coord<T> {
-    type Output = Self;
-
-    fn mul(self, rhs: f64) -> Self {
-        Coord {
-            coord: self.coord * rhs,
-            space: self.space,
-        }
-    }
-}
-
-struct WatermarkPipeline {
-    bind_group_layout: wgpu::BindGroupLayout,
-    render_pipeline: wgpu::RenderPipeline,
-}
-
-impl WatermarkPipeline {
-    pub fn new(device: &wgpu::Device) -> Self {
-        let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-            label: Some("WatermarkBindGroupLayout"),
-            entries: &[
-                wgpu::BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Texture {
-                        sample_type: wgpu::TextureSampleType::Float { filterable: true },
-                        view_dimension: wgpu::TextureViewDimension::D2,
-                        multisampled: false,
-                    },
-                    count: None,
-                },
-                wgpu::BindGroupLayoutEntry {
-                    binding: 1,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::Filtering),
-                    count: None,
-                },
-                wgpu::BindGroupLayoutEntry {
-                    binding: 2,
-                    visibility: wgpu::ShaderStages::VERTEX_FRAGMENT,
-                    ty: wgpu::BindingType::Buffer {
-                        ty: wgpu::BufferBindingType::Uniform,
-                        has_dynamic_offset: false,
-                        min_binding_size: None,
-                    },
-                    count: None,
-                },
-            ],
-        });
-
-        let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
-            label: Some("WatermarkShader"),
-            source: wgpu::ShaderSource::Wgsl(include_str!("shaders/watermark.wgsl").into()),
-        });
-
-        let empty_constants: HashMap<String, f64> = HashMap::new();
-
-        let render_pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-            label: Some("WatermarkPipeline"),
-            layout: Some(
-                &device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-                    label: Some("WatermarkPipelineLayout"),
-                    bind_group_layouts: &[&bind_group_layout],
-                    push_constant_ranges: &[],
-                }),
-            ),
-            vertex: wgpu::VertexState {
-                module: &shader,
-                entry_point: "vs_main",
-                buffers: &[],
-                compilation_options: wgpu::PipelineCompilationOptions {
-                    constants: &empty_constants,
-                    zero_initialize_workgroup_memory: false,
-                    vertex_pulling_transform: false,
-                },
-            },
-            fragment: Some(wgpu::FragmentState {
-                module: &shader,
-                entry_point: "fs_main",
-                targets: &[Some(wgpu::ColorTargetState {
-                    format: wgpu::TextureFormat::Rgba8UnormSrgb,
-                    blend: Some(wgpu::BlendState {
-                        color: wgpu::BlendComponent {
-                            src_factor: wgpu::BlendFactor::One,
-                            dst_factor: wgpu::BlendFactor::OneMinusSrcAlpha,
-                            operation: wgpu::BlendOperation::Add,
-                        },
-                        alpha: wgpu::BlendComponent {
-                            src_factor: wgpu::BlendFactor::One,
-                            dst_factor: wgpu::BlendFactor::OneMinusSrcAlpha,
-                            operation: wgpu::BlendOperation::Add,
-                        },
-                    }),
-                    write_mask: wgpu::ColorWrites::ALL,
-                })],
-                compilation_options: wgpu::PipelineCompilationOptions {
-                    constants: &empty_constants,
-                    zero_initialize_workgroup_memory: false,
-                    vertex_pulling_transform: false,
-                },
-            }),
-            primitive: wgpu::PrimitiveState {
-                topology: wgpu::PrimitiveTopology::TriangleStrip,
-                strip_index_format: None,
-                front_face: wgpu::FrontFace::Ccw,
-                cull_mode: None,
-                unclipped_depth: false,
-                polygon_mode: wgpu::PolygonMode::Fill,
-                conservative: false,
-            },
-            depth_stencil: None,
-            multisample: wgpu::MultisampleState::default(),
-            multiview: None,
-            cache: None,
-        });
-
-        Self {
-            bind_group_layout,
-            render_pipeline,
-        }
-    }
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
-struct WatermarkUniforms {
-    output_size: [f32; 2],
-    watermark_size: [f32; 2],
-    position: [f32; 2],
-    opacity: f32,
-    is_upgraded: f32,
-}
-
-struct ImageBackgroundPipeline {
-    pub bind_group_layout: wgpu::BindGroupLayout,
-    pub render_pipeline: wgpu::RenderPipeline,
-}
-
-#[repr(C)]
-#[derive(Copy, Clone, Pod, Zeroable)]
-struct ImageBackgroundUniforms {
-    output_size: [f32; 2],
-    padding: f32,
-    x_width: f32,
-    y_height: f32,
-    _padding: f32, // For alignment
-}
-
-impl ImageBackgroundPipeline {
-    pub fn new(device: &wgpu::Device) -> Self {
-        let bind_group_layout = device.create_bind_group_layout(&wgpu::BindGroupLayoutDescriptor {
-            label: Some("ImageBackgroundBindGroupLayout"),
-            entries: &[
-                wgpu::BindGroupLayoutEntry {
-                    binding: 0,
-                    visibility: wgpu::ShaderStages::VERTEX_FRAGMENT,
-                    ty: wgpu::BindingType::Buffer {
-                        ty: wgpu::BufferBindingType::Uniform,
-                        has_dynamic_offset: false,
-                        min_binding_size: None,
-                    },
-                    count: None,
-                },
-                wgpu::BindGroupLayoutEntry {
-                    binding: 1,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Texture {
-                        sample_type: wgpu::TextureSampleType::Float { filterable: true },
-                        view_dimension: wgpu::TextureViewDimension::D2,
-                        multisampled: false,
-                    },
-                    count: None,
-                },
-                wgpu::BindGroupLayoutEntry {
-                    binding: 2,
-                    visibility: wgpu::ShaderStages::FRAGMENT,
-                    ty: wgpu::BindingType::Sampler(wgpu::SamplerBindingType::Filtering),
-                    count: None,
-                },
-            ],
-        });
-
-        let shader = device.create_shader_module(wgpu::ShaderModuleDescriptor {
-            label: Some("ImageBackgroundShader"),
-            source: wgpu::ShaderSource::Wgsl(include_str!("shaders/image-background.wgsl").into()),
-        });
-
-        let empty_constants: HashMap<String, f64> = HashMap::new();
-
-        let render_pipeline = device.create_render_pipeline(&wgpu::RenderPipelineDescriptor {
-            label: Some("ImageBackgroundPipeline"),
-            layout: Some(
-                &device.create_pipeline_layout(&wgpu::PipelineLayoutDescriptor {
-                    label: Some("ImageBackgroundPipelineLayout"),
-                    bind_group_layouts: &[&bind_group_layout],
-                    push_constant_ranges: &[],
-                }),
-            ),
-            vertex: wgpu::VertexState {
-                module: &shader,
-                entry_point: "vs_main",
-                buffers: &[],
-                compilation_options: wgpu::PipelineCompilationOptions {
-                    constants: &empty_constants,
-                    zero_initialize_workgroup_memory: false,
-                    vertex_pulling_transform: false,
-                },
-            },
-            fragment: Some(wgpu::FragmentState {
-                module: &shader,
-                entry_point: "fs_main",
-                targets: &[Some(wgpu::ColorTargetState {
-                    format: wgpu::TextureFormat::Rgba8UnormSrgb,
-                    blend: Some(wgpu::BlendState::REPLACE),
-                    write_mask: wgpu::ColorWrites::ALL,
-                })],
-                compilation_options: wgpu::PipelineCompilationOptions {
-                    constants: &empty_constants,
-                    zero_initialize_workgroup_memory: false,
-                    vertex_pulling_transform: false,
-                },
-            }),
-            primitive: wgpu::PrimitiveState {
-                topology: wgpu::PrimitiveTopology::TriangleList,
-                strip_index_format: None,
-                front_face: wgpu::FrontFace::Ccw,
-                cull_mode: Some(wgpu::Face::Back),
-                unclipped_depth: false,
-                polygon_mode: wgpu::PolygonMode::Fill,
-                conservative: false,
-            },
-            depth_stencil: None,
-            multisample: wgpu::MultisampleState::default(),
-            multiview: None,
-            cache: None,
-        });
-
-        Self {
-            bind_group_layout,
-            render_pipeline,
-        }
-    }
-
-    pub fn bind_group(
-        &self,
-        device: &wgpu::Device,
-        uniforms: &wgpu::Buffer,
-        texture: &wgpu::TextureView,
-    ) -> wgpu::BindGroup {
-        let sampler = device.create_sampler(&wgpu::SamplerDescriptor {
-            address_mode_u: wgpu::AddressMode::ClampToEdge,
-            address_mode_v: wgpu::AddressMode::ClampToEdge,
-            address_mode_w: wgpu::AddressMode::ClampToEdge,
-            mag_filter: wgpu::FilterMode::Linear,
-            min_filter: wgpu::FilterMode::Linear,
-            mipmap_filter: wgpu::FilterMode::Linear,
-            ..Default::default()
-        });
-
-        device.create_bind_group(&wgpu::BindGroupDescriptor {
-            label: Some("ImageBackgroundBindGroup"),
-            layout: &self.bind_group_layout,
-            entries: &[
-                wgpu::BindGroupEntry {
-                    binding: 0,
-                    resource: uniforms.as_entire_binding(),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 1,
-                    resource: wgpu::BindingResource::TextureView(texture),
-                },
-                wgpu::BindGroupEntry {
-                    binding: 2,
-                    resource: wgpu::BindingResource::Sampler(&sampler),
-                },
-            ],
-        })
-    }
 }
